+++ sparta_news/post/models.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from django.db import models
+
+class spartanews(models.Model):
+    title = models.CharField(max_length=100)
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.title
