+++ sparta_news/sparta_news/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'accounts',
     'rest_framework_simplejwt.token_blacklist',
+    'post',
 ]
 
 MIDDLEWARE = [
