+++ sparta_news/sparta_news/urls.py
@@ -20,4 +20,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/accounts/', include("accounts.urls")),
+    path('api/post/', include('post.urls')),
 ]
